feat: created components from AppBar and Drawer
- Made custom appbar and drawer components.
- Reused those components in the different types of Layout.
- Created Temporal Drawer Layout using both AppBar and Drawer.
