+++ src/components/Layout/NotLogedInAppBar.tsx
@@ -1,10 +1,8 @@
-import { AppBar, Box, Container, Toolbar } from '@mui/material';
+import { Container } from '@mui/material';
 import { Outlet } from 'react-router-dom';
-import { frontUrls } from '../../data/Urls';
+import { CustomAppBar } from './AppBar';
 
 export function Layout() {
-
-    let qrsecLogo = '/QRSec logo.svg';
 
     return (
         <div id='layout' >
@@ -12,17 +10,11 @@ export function Layout() {
             
             {/* App Bar */}
             <div>
-                <AppBar id='navbar' enableColorOnDark >
-                    <Toolbar>
-                        <Box display='flex' alignItems='center' gap={2}>
-                            <a href={ frontUrls.wholeBase } ><img id='navbar-logo' src={ qrsecLogo } alt='QRSec Logo' /></a>
-                        </Box>
-                    </Toolbar>
-                </AppBar>
+                <CustomAppBar />
             </div>
 
             {/* Main Content */}
-            <div id="content-with-navbar">
+            <div id='content-with-navbar'>
                 <Container maxWidth='sm' >
                     <Outlet />
                 </Container>
